ref: Move form for add new address to separate component
